ui: configuration: add panel for viewing and editing acl entries
using the new generic acl viewing and editing components from
proxmox-yew-comp make it possible for users to edit acl entries

note that the path selector is based on the current definition of
acceptable paths in the `AccessControlConfig`, since those are the
only valid paths. however, the convenience could be improved by, for
example, adding entries for specific resources. however, the acl paths
here don't seem fixed yet, so leave this as-is for now.

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20250828105912.294887-12-s.sterz@proxmox.com