server: add convenience helper to fetch results from many remote nodes in parallel
Implementing this over and over again is a lot of work; many parts of
PDM need to fetch some kind of resource from all remotes and all nodes
of these remotes.

The general approach is the same as for remote task fetching (JoinSet
and semaphores to control concurrency), it will be refactored in a later
commit to use this new helper.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>