server: add api for getting available updates/changelogs for remote nodes
This adds new APIs for update management:

    GET /pve/remotes/{remote}/nodes/{node}/apt/changelog
      -> get package changelog
    GET /pve/remotes/{remote}/nodes/{node}/apt/update
      -> get list of updatable packages
    POST /pve/remotes/{remote}/nodes/{node}/apt/update
      -> refresh APT database

At this time these just pass the call through to PVE with no caching
involved on the PDM side. This should be fine for this API, but once
we have an API for 'give me a view of ALL available remote updates',
we need to introduce a cache that is periodically refreshed.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Tested-by: Stefan Hanreich <s.hanreich@proxmox.com>
Reviewed-by: Stefan Hanreich <s.hanreich@proxmox.com>