server: add locked sdn client helpers
Add a new client that represents a remote with a locked SDN
configuration. It works by creating a new PveClient and then locking
the SDN configuration via the client. It ensures that, while the lock
is held, all methods are called with the proper lock secret. It also
provides helpers for applying / rollbacking the configuration and
releasing the lock.

Additionally, a collection type is introduced, that can hold multiple
locked SDN clients. It provides a helper for executing SDN API calls
across multiple remotes and rolling back the SDN configuration of all
remotes if any API command on any remote fails. This client will be
used for making changes across all remotes in PDM worker tasks.

After applying, we cannot rollback the configuration - so the
collection only logs any errors - waits for the execution to finish on
all remotes and then returns a Result indicating whether the operation
was successful on *all* remotes.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Reviewed-by: Dominik Csapak <d.csapak@proxmox.com>
Tested-by: Dominik Csapak <d.csapak@proxmox.com>