ui: sdn: add view for showing evpn zones
This component shows the current state of all EVPN zones across all
remotes. It shows which zones are configured on a remote and which
VNets it contains. It also shows which remote VNets get imported via
the 'Import Route Target' setting in the EVPN zone. While the ZoneTree
component shows the contents of an IP-VRF merged across all remotes,
the RemoteTree shows the contents of a specific zone on a specific
remote, including all imported VNets.

Similar to the ZoneTree component, this component operates under the
assumption that ASNs are not reused across different remotes, unless
those zones are actually interconnected. For example: One zone imports
a specific ASN:VNI route target. If two zones with the same ASN:VNI
combination exist, but only one of them is connected to the zone, it
will still show the VNets of *both* zones as imported, since it merges
based on ASN:VNI.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Reviewed-by: Dominik Csapak <d.csapak@proxmox.com>
Tested-by: Dominik Csapak <d.csapak@proxmox.com>