ui: sdn: add evpn overview panel
This panel shows an overview of the state of SDN EVPN Zones across
multiple remotes. It includes two different views: a per-remote and a
per-VRF view. For details on the specific views consult the respective
commits. It handles the fetching of data and passing them to the
specific child components and it also handles the dialogues for
creating new EVPN entities (zones, vnets).

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Reviewed-by: Dominik Csapak <d.csapak@proxmox.com>
Tested-by: Dominik Csapak <d.csapak@proxmox.com>