ui: status icons: replace background color with text-shadow
The background color resulted in a visible block, especially when one
hovered a row—as otherwise the used color normally matched already the
color from the panels background anyway, so it was less noticeable.

As font awesome icons are–as the name suggests–fonts, we can simply
use a text-shadow, as that actually follows the icons shape.  Keep
using the background color and use a 2px blur without offset.
The effect is slightly less noticeable compared to setting a
background color, but it still provides some distinction between
background icon and status icon and more importantly, does not looks
so ugly if the overall background color of the panel/row changes
(due to hover, focus, ...).

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>