ui: dashboard: rework default refresh and max-age interval values
Lower the default max age to 30s, which is a somewhat good default for when
having the dashboard open for a prolonged time. Note that I plan to
special case the first load with a higher value in a future commit to
avoid having to compromise to much here.

Actually lower the default refresh interval, as querying itself is
cheap, at least if the backend's cache age is new enough, and doing
that  more often while keeping a higher max-age ensures we profit from
any actual data update, be it those from the backend's polling task or
those from force-reloads from other updates.

Also add some rationale to the doc comments, in the end this is
naturally still chosen by gut-feeling and cannot really be derived
with a objective and logical deduction.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>