ui: dashboard: use high max-age for initial load
Use the same interval as the backend's cache polling task to try to
ensure we get data shown to the user fast, as otherwise it's rather
bad UX if (some) remotes are rather slow.

It's planned to trigger a update with the actual configured polling
interval directly afterwards in a future commit, but this change
itself makes sense on it's own.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>