ui: dashboard: make reload button do a reload with low max-age
Which forces a reload, but not make this a hard cut by using zero
max-age, as that doesn't ensures real time data either (it will be
always some 10 to 100ms old) but rather allow it to be at max 3s old.
This ensures that high-frequent reload button clicks (also from other
users) do not result in a overload/DOS of the remotes.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>