ui: dashboard: default to showing tasks from last 2 days (48h)
As mentioned in the doc comment, this ensures that one sees the full
last day even if one opens the dashboard rather late in the current
day while not having an overly huge time window.

So sure, it's still a subjective change and one should not take that
argument to literal, but personally I use last two days as default for
all my metrics/dashboards I manage (like e.g. grafana or home
assistant), as most often it really does provide the best balance for
me, and we can still change the default if some better argument can be
provided.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>