ui: top entities: make usage threshold a parameter and increase it for memory
The threshold here is not a hard switching point but rather used to
calculate a gradient depending on usage in the oklab color space. As
different metrics have different semantics w.r.t. specific load usage
percentages avoid using a hard coded threshold for all types and
rather enforce that the caller defines this. Use a higher one for
memory, as that is most often fine to (almost) fully use.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>