ui: dashboard: correctly differentiate the timer and on-click reload events
We recently switch this to do a force reload when clicking the refresh
button, but it went unnoticed that the periodic reload is also handled
by the dashboard status bar (a bit odd, but why not) and uses the same
signal callback either way. So expand the callback signature by a bool
parameter to allow signaling a force reload (button clicked) from a
regular reload (update intervall expired).

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>