server: api: add resource-type parameter to list_resources
Add a new parameter to the list_resources call that filters entities
based on their type. While this should already be possible with the
search function as well, filtering directly by enum is more efficient
than building the search filters and then doing matching based on
strings. The type parameter has precedence, so any search filters will
be applied only on entities that have the type specified in
resource-type.

In pdm-client add a separate method to not pollute existing call sites
with unnecessary parameters. In the future it might make sense to add
another method that takes all possible combinations of parameters.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250909155423.526917-6-s.hanreich@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>