ui: add sdn status report to dashboard
This also includes support for external links, searching and
navigating to the dedicated SDN overview. For now, add it to the task
summary row, since there are issues with only showing one element in a
row due to the CSS rules. While this breaks a little bit with the
current grouping, the widget is quite small so it would look weird in
a single row and we can always decide to move it around as soon as we
add more elements to the dashboard.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250909155423.526917-7-s.hanreich@proxmox.com
 [TL: fix fallout from changes that got applied since this series was
  submitted mostly related to self.status now being an Option. Inline
  extracting the SDN zone status into the create_sdn_panel, to avoid
  some odd shuffling of data in circles]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>