ui: main/main menu: fix handling of remote list cache
Commit
0f84394 (ui: main menu: use initial remote list value from context)
tried to handle the initial value of the RemoteList context.

While this fixed one race (loading of the remote list finished before
the main menu was created) it introduced another race: if the load now
takes longer than the creation, we'll overwrite the PersistentState
(browser local storage) once with an empty list and once with the
correct one. This leads to a refresh not keeping the route since at that
time the route does not exists (since the remotelist is empty)

To properly fix this, we have to couple the remote list update with the
cache update, as that is the only point in time where we want to do
that.

For this we have to move the cache update logic to the main entrypoint
(where we update the RemoteList context too) and give the main menu the
list as a a property (from the cache). This way the main menu always has
the view from the cache and this get's updated on every successful api
call to update the remote list.

We still keep the remotelist context around, since other panels deeper
might need it to query information (e.g. the RemoteSelector)

Fixes: 0f84394 (ui: main menu: use initial remote list value from context)
Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>