server: api: resources: improve search when searching for remotes
currently, 'remote:foo' marks the filter as 'remote_only' so
subresources will never be returned. This leads to some awkward searches
that intuitively should work, for example:

'+remote:foo type:qemu' to look for all vms on a specific remote

by making the 'remote' category not 'remote only', and giving the remote
name to the resource filter, we can make that query above work without
problems.

This makes the search with remote names much more intuitive

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20250910132331.506716-2-d.csapak@proxmox.com