ui: css: limit max-height of dropdown picker
Mostly relevant for fields with lots of options, like the timezone
selection or language selection fields, as there the dropdown picker
was shown over the whole view port height, which is not really useful
with so many options and looks rather odd, even impacts UX a bit in a
negative way.

As stop-gap limit the max height to half of the view port minus some
head-room for the field itself to avoid that the picker partially
overlaps the field on certain smaller view port heights, i.e. 45% of
the view port height works out OK here.

Only do this as override here as this is not the best possible
solution, Dominik wants to take a closer look on this when there is a
bit more time, potentially moving the behavior closer to what ExtJS
did, i.e. either show it just below or just above the field, but never
on both sides, but still keeping the filter field inside the dropdown
picker, which has some better UX semantics.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>