ui: remote: check connection for PBS remotes in remote add wizard
Adds the TLS connection check for PBS remotes currently not
implemented in the wizard.

Uses the PDM client to check the connection to a remote PBS instance
with the provided hostname (port) and fingerprint.

Since the pdm client will use the implementation based on the provided
remote type, pass it along from the wizard page props, which is
extended to provide this information.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Reviewed-by: Lukas Wagner <l.wagner@proxmox.com>
Tested-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20250922110958.369653-6-c.ebner@proxmox.com