ui: make the user tab reload when re-opened
otherwise the view would just be cached when switching between tabs.
this can lead to strange outcomes. for example, when a realm sync job
is run, switching back to the list of users still shows the old state.
by updating the key if the state of the tab changes to `visible` the
component is re-rendered and, thus, the load call in the component is
issued again.

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20250922150519.399573-14-s.sterz@proxmox.com