ui: rework status row/meter helpers
By using and returning the already existing MeterLabel from yew-comp, we
can simplify our helpers here. This makes the individual helpers in the
pve panels unnecessary, since we can just use the ones from renderer.

The biggest difference here is that we now use the 'fa-icon' classes
instead of giving a whole 'Fa' struct.

Visually, the result should be identical to before this change.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20250923095124.1679038-11-d.csapak@proxmox.com