ui: pbs: convert snapshot list to a snapshot tree
this is a 'light' version of the tree we have in PBS, only relying on
the snapshot list api call, with less columns.

Since we lack the namespace/group api calls at the moment, just show the
root namespace for now.

It still uses the iterative 'application/json-seq' api behavior to laod
the snapshot list while it's still loading, but instead of always
batching 32 snapshots, batch as many as we can within 250ms.

This should reduce the load of the tree when inserting/resorting but
still give a good indicator that something is happening.

Also adds a loading bar as long as the streaming is ongoing.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20250926072749.560801-9-d.csapak@proxmox.com