ui: pbs: use new pbs panels/overview
this makes use of the newly introduced remote overview panel, the
datastore tree and the datastore panels for PBS remotes.

The intent is for it to look similar to the PVE remote view, but i opted
for making the left hand side (with the datastore tree) narrower, so we
have more space for showing content.

We can do this, because there is less information to be shown per
datastore (name only) than on PVE (tags, actions).

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20250926072749.560801-12-d.csapak@proxmox.com