pbs api: assign ACL with admin role on / for newly created token
While we can just disable privilege separation for the created token
with PVE remotes to get full admin rights, that feature is not
available for our other projects due to trying for a safer design
there as it's more obvious what the permissions are from just checking
the ACL config tree.

That's why we explicitly configure an ACL for the newly created token,
giving it the permissions of the admin role (all privs available
besides some root-only ones) on the / root ACL object path.

As the token will still be one of the configured user, it's real
permissions are still always reduced to the ones of that user at max.

The implementation is pretty simple, albeit we should probably move
out more API types to pbs_api_types for better and safer reuse, even
if we then still want to impl a local type here to decouple them for
different (not fully compatible) API type versions, but that's a whole
other story.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>