ui: dashboard: allow add PBS remotes via dashboard
Extend the current add button to a dropdown menu so it is possible to
add PVE as well as PBS remotes.

Instead of using a boolean to show/hide the wizard, set the remote type
as option, not showing the wizard if none is set, otherwise showing the
wizard for the corresponding type.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250926114535.413361-1-c.ebner@proxmox.com