pbs client: add support to list datastore namespaces
Just a thin wrapper around querying the remote API with using the
existing PBS API types as much as possible.

Single oddity is that we use "datastore" as member name in the struct
that defines the required parameters for this call, while PBS uses
"store" for that. But as the PDM API router currently has the
datastore endpoints under "datastore" and we cannot rename variables
from the API path into a flattened type member IIRC we need to adapt
to this at the pbs client type level.

We should really look into more synergies with the PBS (CLI) client,
as we probably can factor out a lot of common stuff into a pbs client
crate living in the proxmox workspace.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>