d/control: depend on proxmox-login 1.0.2
proxmox-login version 1.0.2 contains commit b0d98294
("login: use `ticket` if both it and `ticket_info` are provided") of
proxmox repo, fixing ticket parsing issues for some versions of
Proxmox Backup Server 3. This therefore lead to the API token
creation failing with and authentication error when adding the PBS as
remote.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20251006113311.568044-2-c.ebner@proxmox.com