fix #6901: api: remove `node` reference from templated privilege path
If a non root user tried to view the overview of a PBS, it was
showing "403: permission check failed" error. This occured because the
privilege path included the "node" object which is neither accepted
as a parameter in the endpoint nor passed from the UI.

To fix the issue removed the "node" reference. Now if the user has
at least the `Resource.Audit` permission, they can view PBS status under
the overview panel.

Signed-off-by: Shan Shaji <s.shaji@proxmox.com>
Reviewed-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20251014085651.73407-3-s.shaji@proxmox.com