api: resources: new transient type for remote resource gathering
Introduce a new transient type to be returned by the get resources
implementation helper, which also includes the full remote object.
Since the remote object can store secret information, it must be
avoided to return it in the API.

Requires now the mapping of the transient type to the final
`RemoteResources` for the API response at the cost of additional
looping over the list of remote resources.

In preparation for adding search capabilities to filter remote
resources based on remote type.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20251021111129.294349-5-c.ebner@proxmox.com