server: api: add remote-type search category for resources
Extend the current search capability for resources by adding the
`RemoteType` search category, allowing to selectively filter remotes
of type PVE and PBS.

Since the current remote filtering is only applied in case of remotes
only search, add an additional helper to pre-filter the to be fetched
resources by the remote type.

With this it is now possible to search, e.g. `remote-type:pbs` to
only get resources from PBS remotes, with the intention to provide
a search shortcut to show PBS remotes when clicking in the to be
added PBS status dashboard panel.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20251021111129.294349-6-c.ebner@proxmox.com