ui: dashboard: reimplement node status panel as dedicated component
Reimplement the current node status panel as a dedicated component
for better encapsulation and input data handling, analogous to how
guest status or SDN status panels are handled.

Instead of passing in the whole status, only pass the state required
for the component to be rendered.

Already take into account that this will be used to render the PBS
node status as well.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20251021111129.294349-9-c.ebner@proxmox.com