pdm-api-types: extend resources by properties string generator method
Generates a list of comma separated values of properties to be
matched by the search functionality.

This will allow to define several properties on a resource which one
can match against. The current usecase is to match PBS datastores
based on properties such as if it is a removable datastore, if the
datastore has high usage or if the datastore is backed by S3.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20251021111129.294349-18-c.ebner@proxmox.com