api: add API for retrieving/refreshing the remote update summary
This commit adds two new endpoints, namely
  GET  /remote-updates/summary
  POST /remote-updates/refresh

The first one is used to retrieve the update summary (the data is taken
from the cache), the second one can be used to proactively refresh the
summary in the cache (starts a worker task, since this could take a
while). Note that we only retrieve the up-to-date list of packages from
the remote, but do *not* trigger an `apt update` right now. Could make
sense to do the latter as well, but then we probably should
stream/forward the task logs for the update task from the node to the
native PDM task; something we can rather implement later.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Reviewed-by: Shannon Sterz <s.sterz@proxmox.com>
Tested-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20251023124420.244585-4-l.wagner@proxmox.com