api types: increase high-usage threshold from 75% to 80%
Still not really ideal, relative usages can often be rather wrong or
useless when the total is huge, the usage significant due to the
initial usage being not so small but the growth rate actually rather
slow, as then such a threshold can be triggered relatively early even
though there is a long time left until the usage becomes actually
problematic.

This should be probably either split into two levels (notice-like and
criticial-like), or take into account recent'ish growth, or the
allowed remainding space be capped with a total, not just a relative
amount, so that big setups do not get matched against, or maybe
reworked completely – but it naturally can be nice to know if
something fills up, so the essence of this is an important metric.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>