ui: dashboard: refactor guest panel creation to its own module
so we can more easily reuse it. For this, also make the 'create_title_with_icon'
a freestanding function that is public so we can reuse it outside the
dashboard struct.

Also add the functionality of not passing a guest_type, so that we can
create a panel that includes info for all guests, regardless of type.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251031124822.2739685-2-d.csapak@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>