ui: dashboard: status row: make loading less jarring
by tracking the loading state not only via the last_refresh property,
but also from a reload message to a changing last_refresh value.

This means it now always shows the last refresh time if there was one
and only rotates the icon in that case.

This is much less jarring that a split second change of the text from
a date to 'now refreshing' back to a finished date.

While at it, remove the unused 'loading' property of the PdmDashboard
struct.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251031124822.2739685-16-d.csapak@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>