ui: dashboard: implement 'View'
this is a more generalized version of our `Dashboard`, which can be
configured with a `ViewTemplate` that is completely serializable, so
we're able to dynamically configure it (e.g. in the future over the api)

An example of such a configuration could be, e.g.:

{
    "layout": {
        "layout-type": "rows",
        "rows": [
            [
                {
                    "widget-type": "remotes",
                    "show-wizard": true
                }
            ],
            [
                {
                    "widget-type": "leaderboard",
                    "leaderboard-type": "guest-cpu"
                }
            ]
        ]
    }
}

Implemented are all widget our Dashboard holds, but this is easily
extendable.

We abstract the 'RowView' away in it's own module, so we can easily add
a new layout types.

To avoid some unnecessary cloning, use a 'SharedState' for the load
results, that way only Rc's will be cloned into the render callback.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251031124822.2739685-18-d.csapak@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>