ui: dashboard guest panel: split imports at module-level
And improve grouping to go from lower-level dependencies to higher
level ones (that often depend itself on the lower-level ones).

The module-level split makes merging patches easier as it strongly
reduces the chance for triggering reshuffling of complex multi-line
imports; and if it still happens with module-level imports it's often
a indicator of bad export hygiene, i.e. where the dependency should
better group its imports in submodules, or provide a prelude or the
like.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>