api: fail with HTTP 403 Forbidden if user does have access to resource
As discussed in another patch series [1], if a user does not have
enough permissions to see a resource, we should fail with 403
Forbidden instead of 401 Unauthorized. As the latter (401) means that
there was no valid ticket or API token sent along with the request,
whereas the former (403) is for requests which have valid
authentication, but where the user or token is missing the required
permissions. This follows the semantics used in our other projects,
like PVE.

The UI relies on returning the correct error codes, since a 401 will
redirect the user to the login mask.

[1] https://lore.proxmox.com/pdm-devel/DDJLG0Y8J5CP.1NMNIMD5ZEYZE@proxmox.com/

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20251105081254.40625-2-l.wagner@proxmox.com
 [TL: slightly extend reasoning for why 403 should be used]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>